Separated admin and user views and permissions. Stored is_staff state in local storage and passed down as prop after parsing bool. Displayed all users in a list when an admin is logged in.
