+++ src/managers/users.js
@@ -1,5 +1,12 @@
-export const getUsers = () => {
-    return fetch("http://localhost:8000/users")
+export const getAllAuthors = () => {
+    return fetch(`http://localhost:8000/authors`, {
+        method: "GET",
+        headers: {
+            "Content-Type": "application/json",
+            "Accept": "application/json",
+            "Authorization": `Token ${localStorage.getItem("auth_token")}`
+        }
+    })
         .then(res => res.json())
 }
